requirements.txt
telethon
python-dateutil
pandas
pytesseract
python-magic
tqdm
python-dotenv